# Optimize Variables and switches + Add Tests