--- a/src/game_battler.cpp
+++ b/src/game_battler.cpp
@@ -366,7 +366,7 @@ bool Game_Battler::UseSkill(int skill_id, const Game_Battler* source) {
 		was_used = true;
 	} else if (skill->type == RPG::Skill::Type_switch) {
 		Game_System::SePlay(skill->sound_effect);
-		Game_Switches.Set(skill->switch_id, true);
+		Main_Data::game_switches->Set(skill->switch_id, true);
 		was_used = true;
 	}