--- a/src/game_character.cpp
+++ b/src/game_character.cpp
@@ -363,12 +363,12 @@ void Game_Character::UpdateMoveRoute(int32_t& current_index, const RPG::MoveRout
 				SetMoveFrequency(max(GetMoveFrequency() - 1, 1));
 				break;
 			case RPG::MoveCommand::Code::switch_on: // Parameter A: Switch to turn on
-				Game_Switches.Set(move_command.parameter_a, true);
+				Main_Data::game_switches->Set(move_command.parameter_a, true);
 				Game_Map::SetNeedRefresh(Game_Map::Refresh_All);
 				Game_Map::Refresh();
 				break;
 			case RPG::MoveCommand::Code::switch_off: // Parameter A: Switch to turn off
-				Game_Switches.Set(move_command.parameter_a, false);
+				Main_Data::game_switches->Set(move_command.parameter_a, false);
 				Game_Map::SetNeedRefresh(Game_Map::Refresh_All);
 				Game_Map::Refresh();
 				break;