--- a/src/game_event.cpp
+++ b/src/game_event.cpp
@@ -216,34 +216,34 @@ bool Game_Event::AreConditionsMet(const RPG::EventPage& page) {
 
 	// Variable
 	if (Player::IsRPG2k()) {
-		if (page.condition.flags.variable && !(Game_Variables.Get(page.condition.variable_id) >= page.condition.variable_value)) {
+		if (page.condition.flags.variable && !(Main_Data::game_variables->Get(page.condition.variable_id) >= page.condition.variable_value)) {
 			return false;
 		}
 	} else {
 		if (page.condition.flags.variable) {
 			switch (page.condition.compare_operator) {
 			case 0: // ==
-				if (!(Game_Variables.Get(page.condition.variable_id) == page.condition.variable_value))
+				if (!(Main_Data::game_variables->Get(page.condition.variable_id) == page.condition.variable_value))
 					return false;
 				break;
 			case 1: // >=
-				if (!(Game_Variables.Get(page.condition.variable_id) >= page.condition.variable_value))
+				if (!(Main_Data::game_variables->Get(page.condition.variable_id) >= page.condition.variable_value))
 					return false;
 				break;
 			case 2: // <=
-				if (!(Game_Variables.Get(page.condition.variable_id) <= page.condition.variable_value))
+				if (!(Main_Data::game_variables->Get(page.condition.variable_id) <= page.condition.variable_value))
 					return false;
 				break;
 			case 3: // >
-				if (!(Game_Variables.Get(page.condition.variable_id) > page.condition.variable_value))
+				if (!(Main_Data::game_variables->Get(page.condition.variable_id) > page.condition.variable_value))
 					return false;
 				break;
 			case 4: // <
-				if (!(Game_Variables.Get(page.condition.variable_id) < page.condition.variable_value))
+				if (!(Main_Data::game_variables->Get(page.condition.variable_id) < page.condition.variable_value))
 					return false;
 				break;
 			case 5: // !=
-				if (!(Game_Variables.Get(page.condition.variable_id) != page.condition.variable_value))
+				if (!(Main_Data::game_variables->Get(page.condition.variable_id) != page.condition.variable_value))
 					return false;
 				break;
 			}