--- a/src/game_interpreter_battle.cpp
+++ b/src/game_interpreter_battle.cpp
@@ -174,7 +174,7 @@ bool Game_Interpreter_Battle::CommandChangeMonsterHP(RPG::EventCommand const& co
 		change = com.parameters[3];
 	    break;
 	case 1:
-		change = Game_Variables.Get(com.parameters[3]);
+		change = Main_Data::game_variables->Get(com.parameters[3]);
 	    break;
 	case 2:
 		change = com.parameters[3] * hp / 100;
@@ -206,7 +206,7 @@ bool Game_Interpreter_Battle::CommandChangeMonsterMP(RPG::EventCommand const& co
 		change = com.parameters[3];
 	    break;
 	case 1:
-		change = Game_Variables.Get(com.parameters[3]);
+		change = Main_Data::game_variables->Get(com.parameters[3]);
 	    break;
 	}
 
@@ -319,11 +319,11 @@ bool Game_Interpreter_Battle::CommandConditionalBranchBattle(RPG::EventCommand c
 			break;
 		case 1:
 			// Variable
-			value1 = Game_Variables.Get(com.parameters[1]);
+			value1 = Main_Data::game_variables->Get(com.parameters[1]);
 			if (com.parameters[2] == 0) {
 				value2 = com.parameters[3];
 			} else {
-				value2 = Game_Variables.Get(com.parameters[3]);
+				value2 = Main_Data::game_variables->Get(com.parameters[3]);
 			}
 			switch (com.parameters[4]) {
 				case 0: