--- a/src/game_player.cpp
+++ b/src/game_player.cpp
@@ -69,7 +69,7 @@ void Game_Player::ReserveTeleport(const RPG::SaveTarget& target) {
 	ReserveTeleport(map_id, target.map_x, target.map_y, Down, TeleportTarget::eSkillTeleport);
 
 	if (target.switch_on) {
-		Game_Switches.Set(target.switch_id, true);
+		Main_Data::game_switches->Set(target.switch_id, true);
 		Game_Map::SetNeedRefresh(Game_Map::Refresh_All);
 	}