… CMakeLists.txt
… bench/switches.cpp
… bench/variables.cpp
… src/game_switches.cpp
… src/game_switches.h
… src/game_variables.cpp
… src/game_variables.h
… src/main_data.cpp
… src/main_data.h
… src/player.cpp
… src/scene_debug.cpp
… src/scene_item.cpp
… src/scene_save.cpp
… src/window_message.cpp
… src/window_varlist.cpp
… tests/switches.cpp
… tests/variables.cpp