Optimizes switch and variable access for logic heavy games. Also cleanups singletons and adds unit tests.

Master:
```
----------------------------------------------------------
Benchmark                   Time           CPU Iterations
----------------------------------------------------------
BM_SwitchGet                4 ns          4 ns  189047081
BM_SwitchSet                5 ns          5 ns  136290582
BM_SwitchFlip               7 ns          7 ns  103273384
BM_SwitchSetRange        5009 ns       5009 ns     141417
BM_SwitchFlipRange       6717 ns       6717 ns     105261

------------------------------------------------------------
Benchmark                     Time           CPU Iterations
------------------------------------------------------------
BM_VariableGet                3 ns          3 ns  234503273
BM_VariableSet                4 ns          4 ns  166569307
BM_VariableAdd                7 ns          7 ns  102786779
BM_VariableSub                7 ns          7 ns  100826372
BM_VariableMult               7 ns          7 ns  100840661
BM_VariableDiv                9 ns          8 ns   81506066
BM_VariableMod                8 ns          8 ns   82874145
BM_VariableSetRange        4133 ns       4104 ns     170068
BM_VariableAddRange        6829 ns       6783 ns     104131
BM_VariableSubRange        7187 ns       7129 ns     100112
BM_VariableMultRange       7201 ns       7140 ns      97368
BM_VariableDivRange        9557 ns       9480 ns      73299
BM_VariableModRange        8700 ns       8635 ns      79854
```

This PR:
```
----------------------------------------------------------
Benchmark                   Time           CPU Iterations
----------------------------------------------------------
BM_SwitchGet                1 ns          1 ns  493785383
BM_SwitchSet                5 ns          5 ns  136517949
BM_SwitchFlip               5 ns          5 ns  136289613
BM_SwitchSetRange        1282 ns       1282 ns     551636
BM_SwitchFlipRange       1251 ns       1251 ns     547390


-----------------------------------------------------------------------
Benchmark                     Time           CPU Iterations
------------------------------------------------------------
BM_VariableGet                3 ns          3 ns  267765975
BM_VariableSet                4 ns          4 ns  158529616
BM_VariableAdd                4 ns          4 ns  161540747
BM_VariableSub                5 ns          5 ns  150471152
BM_VariableMult               6 ns          6 ns  130454647
BM_VariableDiv                8 ns          8 ns   94011699
BM_VariableMod                6 ns          6 ns  107436757
BM_VariableSetRange         916 ns        916 ns     757338
BM_VariableAddRange         687 ns        687 ns    1026075
BM_VariableSubRange         618 ns        618 ns    1092798
BM_VariableMultRange        687 ns        687 ns    1025344
BM_VariableDivRange        2366 ns       2366 ns     291713
BM_VariableModRange        2354 ns       2352 ns     294594
```