@fmatthew5876 2019-12-15
Here are some improvements in Frozen Triggers.

This profile was taken for 10 seconds at the beginning of the game with the blowing snow.

Master
![fzold](https://user-images.githubusercontent.com/1198466/70856634-5d7f7b80-1eae-11ea-82fc-033c6d765321.png)

This PR
![fznew](https://user-images.githubusercontent.com/1198466/70856633-5d7f7b80-1eae-11ea-92fd-ad332dd3a5f2.png)