@Ghabry 2019-12-15
Wow this is some significant improvement.
Also that VTune is cool, didn't knew it exists for Linux :o. Looks like the Visual Studio Profiler, now I don't need to boot my Windows VM anymore for this :+1: