@fmatthew5876 2019-12-15
Surprisingly looks like `Game_Map::Refresh()` only takes 0.5% of the time in my profile.