@Ghabry 2019-12-18
Looks also good to me. Only minor thing I see that the switches have a return value of "int". Is that intentional? Shouldn't have side-effects, just looks strange. xD (but if you plan to change this add this to some other PR, don't create PRs for such one line changes pls, dat queue) ;)