@carstene1ns 2019-12-21
This one unfortunately has another long vs. int discrepancy in newlib on 3ds:

```c++
/opt/jenkins/workspace/player-3ds/builds/3ds/../../src/game_variables.cpp: At global scope:
/opt/jenkins/workspace/player-3ds/builds/3ds/../../src/game_variables.cpp:64:23: error: no declaration matches 'Game_Variables::Var_t Game_Variables::SetOp(int, Game_Variables::Var_t, F&&, const char*)'
   64 | Game_Variables::Var_t Game_Variables::SetOp(int variable_id, Var_t value, F&& op, const char* warn) {
      |                       ^~~~~~~~~~~~~~
In file included from /opt/jenkins/workspace/player-3ds/builds/3ds/../../src/game_variables.cpp:19:
/opt/jenkins/workspace/player-3ds/builds/3ds/../../src/game_variables.h:72:7: note: candidate is: 'template<class F> int Game_Variables::SetOp(int, Game_Variables::Var_t, F&&, const char*)'
   72 |   int SetOp(int variable_id, Var_t value, F&& op, const char* warn);
      |       ^~~~~
/opt/jenkins/workspace/player-3ds/builds/3ds/../../src/game_variables.h:29:7: note: 'class Game_Variables' defined here
   29 | class Game_Variables {
      |       ^~~~~~~~~~~~~~
In file included from /opt/jenkins/workspace/player-3ds/builds/3ds/../../src/game_variables.h:23,
                 from /opt/jenkins/workspace/player-3ds/builds/3ds/../../src/game_variables.cpp:19:
/opt/jenkins/workspace/player-3ds/builds/3ds/../../src/game_variables.cpp: In instantiation of 'void Game_Variables::SetOpRange(int, int, Game_Variables::Var_t, F&&, const char*) [with F = long int (&)(long int, long int); Game_Variables::Var_t = long int]':
/opt/jenkins/workspace/player-3ds/builds/3ds/../../src/game_variables.cpp:122:79:   required from here
```